AI avoid front of player fix & more stuff
I believe i managed to fix the behavior for the AI to try and avoid the front of the player, within a minimum angle.
Made adjustments for AI moving in random direction  in ranged and wander behaviors.
Added gizmos drawing for fancyness, it draws the "dangerous" area in front of the player when AI is attacking, as well as the current angle for each AI in relation to the front of the player.
Also added input action context for the AI, to be able to calculate the front and right of the player for itself.