Level gen working & AI now has weapon handler
Got the basics of the level generation working, it is able to successfully generate a fully traversable level of connected "rooms", including being able to teleport player from room to room when walking into a "door".

Also made sure to add weapon handler to the AI prefab now that we have basic shooting working.